requirements.txt
BeautifulSoup4==4.9.1
pytest==6.0.1
azure-cosmos==4.0.0
